Добавил обновление комментария и удаление

+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -11,7 +11,7 @@ public interface CommentRepository extends JpaRepository<Comment, Integer>{
 
     List<Comment> findAllByAdId(Integer adId);
 
-    void deleteByIdAndAdId(long adId, long commentId);
+    void deleteByIdAndAdId(Integer adId, Integer commentId);
 
-    Comment findByIdAndAd_Id(long adId, long commentId);
+    Comment findByIdAndAd_Id(Integer adId, Integer commentId);
 }
